[Bugfix] Import - Utilisation d'entiers longs pour les surfaces de parcelle_info simple
La modification https://github.com/3liz/QgisCadastrePlugin/commit/a6790eb23c97883c0fb58b5cdfe5e3e4fc39208c : Import - Utilisation d'entiers longs pour les surfaces de parcelle_info
qui corrige un souci pour les départements avec des parcelles volumineuses, comme la Guyane ; n'avait pas été appliqué à parcelle_info_simple.

+++ cadastre/scripts/plugin/edigeo_create_table_parcelle_info_simple.sql
@@ -20,7 +20,7 @@ SELECT AddGeometryColumn ( current_schema::text, 'parcelle_info', 'geom', 2154 ,
 
 INSERT INTO [PREFIXE]parcelle_info
 SELECT gp.ogc_fid AS ogc_fid, geo_parcelle, gp.idu AS idu, gp.tex AS tex, gp.geo_section AS geo_section,
-c.tex2 AS nomcommune, c.idu AS codecommune, Cast(ST_Area(gp.geom) AS integer) AS surface_geo, gp.supf AS contenance, 
+c.tex2 AS nomcommune, c.idu AS codecommune, Cast(ST_Area(gp.geom) AS bigint) AS surface_geo, gp.supf AS contenance,
 gp.lot AS lot,
 gp.geom AS geom
 FROM [PREFIXE]geo_parcelle gp
